Do not create a test file where none previously existed

+++ bin/run.sh
@@ -69,8 +69,10 @@ run_tests() {
         [[ -f "$test_file" ]] || test_file="${slug//-/_}_test.sh"
     fi
 
-    contents=$( < "$test_file" )
-    sed 's/(load bats-extra)\.bash/$1/' > "$test_file" <<< "$contents"
+    if [[ -e "$test_file" ]]; then
+        contents=$( < "$test_file" )
+        sed 's/(load bats-extra)\.bash/$1/' > "$test_file" <<< "$contents"
+    fi
 
     echo "Test output:"
